Moving out of beta

+++ app/src/main/kotlin/com/aracroproducts/attentionv2/NetworkUtils.kt
@@ -189,7 +189,7 @@ class ProgressRequestBody(
             while (read != -1) {
                 sink.write(buf, 0, read)
                 uploaded += read
-                progressUpdate?.invoke((uploaded.toDouble() / contentLength).toFloat())
+                progressUpdate?.invoke((uploaded.toFloat() / contentLength))
                 read = image.read(buf)
             }
         } catch (_: IOException) {
